fix(ci): python-coverage-shards — split oversized non-check reviewer settle tests under 1500-line guardrail
